create github component
