+++ src/App.vue
@@ -45,7 +45,8 @@ export default Vue.extend({
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
-  margin-top: 60px;
+  max-width: 1700px;
+  margin: 60px auto 0;
 }
 .footer {
   width: 800px;
